Реализованы тесты ListStorageTest через AbstractStorageTest

+++ src/com/anastas/webapp/storage/ListStorage.java
@@ -38,7 +38,7 @@ public class ListStorage extends AbstractStorage {
 
     @Override
     public void doDelete(Object searchKey) {
-        list.remove((Integer) searchKey);
+        list.remove(((Integer) searchKey).intValue());
     }
 
     @Override
